refactor: Delete api file, move fetchSearchResults function to a separate file
