+++ src/app/[lang]/boxes/page.tsx
@@ -3,7 +3,6 @@ import BoxesPageHeader from '@components/components/BoxesPage/BoxesPageHeader/Bo
 import BoxesSection from '@components/components/BoxesPage/BoxesSection/BoxesSection';
 import Breadcrumbs from '@components/components/Breadcrumbs/Breadcrumbs';
 import { convertToServerLocale } from '@components/helpers/convertToServerLocale';
-import { getBoxes } from '@lib/api-services/api';
 import { fetchBoxes } from '@lib/api-services/fetchBoxes';
 import { getDictionary } from '@lib/utils/dictionary';
